Rename DeleteMusic.js to MusicDelete.js
