+++ client/src/MusicList.js
@@ -1,6 +1,6 @@
 import React, { useState, useEffect } from 'react';
 import { Link } from 'react-router-dom';
-import deleteMusic from "./DeleteMusic"
+import deleteMusic from "./MusicDelete"
 
 export default function MusicList() {
